# Conflicts:
#	charting/model/chart.py
#	examples/charting/output/inflation.png

+++ charting/model/chart.py
@@ -6,6 +6,7 @@ from matplotlib import pyplot as plt
 from matplotlib.axes import Axes
 from matplotlib.ticker import Formatter, Locator, AutoLocator
 from datetime import datetime
+
 from charting.model.style import title_style, source_text_style
 
 
@@ -102,7 +103,7 @@ class Chart(ABC):
         """
         self.ax.tick_params(axis='x', which=which, length=length, width=width, rotation=rotation, pad=pad)
 
-    def configure_y_axis(self, axis_index: int, label: str = None, y_lim: Tuple[int, int] = None,
+    def configure_y_axis(self, axis_index: int, label: str = None, y_lim: Tuple[float, float] = None,
                          minor_formatter: Formatter = None,
                          major_formatter: Formatter = None,
                          minor_locator: Locator = None,
